Factor out county audit wizard, improve stage enum.
